try pods 2.0.0

+++ LinkBlock/LinkInfo/LinkError.m
@@ -10,12 +10,12 @@
 @implementation LinkError
 - (NSString *)description
 {
-    return [NSString stringWithFormat:@"【LinkError】调用方法\"%@\"时需要类型\"%@\"而不是类型\"%@\"，在链条的倒数第\"%lu\"处发生",self.inFunc,self.needClass,self.errorClass,self.throwCount+1];
+    return [NSString stringWithFormat:@"【LinkError】调用方法\"%@\"时需要类型\"%@\"而不是类型\"%@\"，在链条的倒数第\"%lu\"处发生",self.inFunc,self.needClass,self.errorClass,(unsigned long)self.throwCount+1];
 }
 
 - (NSString *)debugDescription
 {
-    return [NSString stringWithFormat:@"LinkError:throwCount=%lu,needClass=%@,errorClass=%@,inFunc=%@",self.throwCount,self.needClass,self.errorClass,self.inFunc];
+    return [NSString stringWithFormat:@"LinkError:throwCount=%lu,needClass=%@,errorClass=%@,inFunc=%@",(unsigned long)self.throwCount,self.needClass,self.errorClass,self.inFunc];
 }
 
 - (instancetype)init
